Add ip_connection.cpp

+++ src/pal/net/ip_connection.hpp
@@ -9,14 +9,14 @@ namespace webserv {
     namespace pal {
         namespace net {
             class ip_connection {
-                int         _port;
+                uint16_t    _port;
                 ip_address  _address;
 
             public:
                 ip_connection(ip_address address, int port);
                 ~ip_connection();
 
-                int get_port();
+                uint16_t get_port();
                 std::string get_address_s();
             };
